don't unlink if link doesn't exist

+++ codedeploy/scripts/stop_unicorn.sh
@@ -1,2 +1,4 @@
 kill -9 $(cat /tmp/unicorn.pid)
-unlink /opt/current-deployment
+if [ -L /opt/current-deployment ]; then
+  unlink /opt/current-deployment
+fi
